+++ index.html
@@ -227,5 +227,5 @@
     </section>
   </template>
   <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </html>
